Make nucleotide counting work when the input source does not provide any information about nucleic acid sequences.
Furthermore, the original implementation appears to have been incorrect
because it did not account for the fact that if there are multiple
chains with the same sequence (such as antiparallel chains), mmCIF file
may write out the sequence only once and use the "pdbx_strand_id"
attribute to indicate the chains with the same sequence. In such cases,
the number of nucleotides must be multiplied accordingly.

Fixes: https://gitlab.cesnet.cz/madcatxster/rednatco/-/issues/3