Replace TTF with WOFF2 for the Roboto font family
WOFF2 is the recommended format to exchange font files on the Web. It is
an open standard and the files are compressed.
This change also fixes a Webpack issue where the .ttf font assets are
not bundled correctly. (We need to handle .ttf files diffently to
make the validation report tool work. This causes the .ttf files to be
processed by two different loaders and they do not get bundled
correctly.)